+++ src/tightdb/column.cpp
@@ -243,6 +243,7 @@ template<class Op> void col_type_deleg(Op& op, ColumnType type)
         case col_type_Int:
         case col_type_Bool:
         case col_type_DateTime:
+        case col_type_Link:
             op.template call<Column>();
             return;
         case col_type_String:
@@ -268,6 +269,8 @@ template<class Op> void col_type_deleg(Op& op, ColumnType type)
             return;
         case col_type_Reserved1:
         case col_type_Reserved4:
+        case col_type_LinkList:
+        case col_type_BackLink:
             break;
     }
     TIGHTDB_ASSERT(false);